pdf-preview now creates by itself one file per PDF page
This is not how I wanted it to work (because from now on, open() is
allowed in the sandbox) but this is at least directly usable.

For the future, I would really prefer to send SVG to a trusted process
doing the file creation and thus remove the open() syscall in the
whitelist.